+++ kaye_engine/cli/claude/vs_code/export.py
@@ -12,7 +12,9 @@ from kaye_engine.cli.claude.marketplace.export import export_marketplace
 from kaye_engine.cli.claude.user_prompt.parser import (
     find_user_system_prompt_file,
 )
-from kaye_engine.cli.claude.user_prompt.export import export_user_system_prompt_file
+from kaye_engine.cli.claude.user_prompt.export import (
+    export_user_system_prompt_file,
+)
 from .settings import update_settings_json
 
 # logger  ######################################################################