+++ kaye_engine/kamilog.py
@@ -65,7 +65,7 @@ __all__ = (
 
 
 # metadata  ####################################################################
-__version__ = "2.9.2"
+__version__ = "2.9.3"
 __author__ = "kamiLeL"
 
 
@@ -181,9 +181,7 @@ class AnsiStyle(Flag):  # =====================================================
             try:
                 style |= cls[name]
             except KeyError:
-                raise ValueError(
-                    "unknown AnsiStyle member {!r}".format(name)
-                )
+                raise ValueError("unknown AnsiStyle member {!r}".format(name))
         return style
 
 
@@ -1320,10 +1318,10 @@ def _set_logger_level(level, *, logger=None, logger_name=None):
 _VERBOSE_FLAG_CHOICES = ("vq", "VQ", "")
 _EXTREME_VERBOSITY = 1_000_000
 
-_STEP_VERBOSE_HELP = "make verbose, each {opt} increase verbosity by 1"
-_STEP_QUIET_HELP = "make quiet, each {opt} decrease verbosity by 1"
-_EXTREMITY_VERBOSE_HELP = "make maximally verbose, via {opt}"
-_EXTREMITY_QUIET_HELP = "make maximally quiet, via {opt}"
+_STEP_VERBOSE_HELP = "increase verbosity by 1 per {opt}"
+_STEP_QUIET_HELP = "decrease verbosity by 1 per {opt}"
+_EXTREMITY_VERBOSE_HELP = "set maximally verbose"
+_EXTREMITY_QUIET_HELP = "set maximally quiet"
 
 
 def _flag_option_strings(short_flag, long_flag):
@@ -1441,7 +1439,7 @@ def add_verbose_arguments(
         action="store_const",
         const=_EXTREME_VERBOSITY,
         default=0,
-        help=_EXTREMITY_VERBOSE_HELP.format(opt="/".join(max_verbose_opts)),
+        help=_EXTREMITY_VERBOSE_HELP,
     )
     parser.add_argument(
         *max_quiet_opts,
@@ -1449,7 +1447,7 @@ def add_verbose_arguments(
         action="store_const",
         const=_EXTREME_VERBOSITY,
         default=0,
-        help=_EXTREMITY_QUIET_HELP.format(opt="/".join(max_quiet_opts)),
+        help=_EXTREMITY_QUIET_HELP,
     )
 
 
@@ -1993,7 +1991,7 @@ def _register_color_parser(cli_subparser):
         metavar="STYLE",
         nargs="+",
         type=_parse_ansi_style,
-        help= "1+ ANSI styles, v.s.",
+        help="1+ ANSI styles, v.s.",
     )
 
     color_parser.set_defaults(func=_color_parser_main)