+++ tests/abbr/prompt-abbr-wrap_test.py
@@ -135,19 +135,62 @@ class TestRuleSymbol:  # =======================================================
 
     # true cases  **************************************************************
     def test_true1(self):
-        ipt = ("", "")
+        ipt = (" ", " ")
         print(ipt)
         assert self.enum.is_satisfied_wrap_rule(*ipt)
 
     def test_true2(self):
-        ipt = ("a", "z")
+        ipt = ("", " ")
         print(ipt)
         assert self.enum.is_satisfied_wrap_rule(*ipt)
 
     def test_true3(self):
-        ipt = ("\n", "\t")
+        ipt = (" ", "")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true4(self):
+        ipt = ("", "")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true5(self):
+        ipt = ("", "!")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true6(self):
+        ipt = ("", ":")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true7(self):
+        ipt = (",", " ")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true8(self):
+        ipt = ("\n", "")
+        print(ipt)
+        assert self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_true9(self):
+        ipt = ("\t", " ")
         print(ipt)
         assert self.enum.is_satisfied_wrap_rule(*ipt)
 
     # false cases  *************************************************************
-    # no false cases
+    def test_false1(self):
+        ipt = ("a", "z")
+        print(ipt)
+        assert not self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_false2(self):
+        ipt = ("a", "")
+        print(ipt)
+        assert not self.enum.is_satisfied_wrap_rule(*ipt)
+
+    def test_false3(self):
+        ipt = (" ", "z")
+        print(ipt)
+        assert not self.enum.is_satisfied_wrap_rule(*ipt)